fix: TW Switch styles bug
